+++ doc/history.rst
@@ -5,6 +5,7 @@ History
 v0.1.0
 -------
 
+  - fix bug where "value" blocked by general "field"
   - *correctOrbit* and *correctOrbitPv* works on new lattice infrastructure.
   - separate orm data from orm measurement
   - add fields (x/y) for BPM/HCOR, change TRIMX/TRIMY to HCOR/VCOR and HFCOR/VFCOR.